Building font bit library for tft display

+++ inc/MS_FONT/BITMAP.h
@@ -3,6 +3,6 @@
 
 #include <sched.h>
 
-uint8_t *bitmap_char(uint8_t CHAR); // Charecter bitmap
+uint16_t *bitmap_char(uint8_t CHAR); // Charecter bitmap
 
 #endif
